pyproject.toml
[tool.poetry]
name = "simtutor-core"
version = "0.1.0"
description = "Simulator-agnostic tutoring backend skeleton for SimTutor Core v0.1"
authors = ["Yu Zhang <yu.zhang.1@tu-clausthal.de>"]
packages = [
  { include = "core" },
  { include = "ports" },
  { include = "adapters" },
  { include = "packs" },
  { include = "tools" },
  { include = "simtutor" },
]
include = ["simtutor/schemas/v1/*.schema.json"]

[tool.poetry.dependencies]
python = "^3.10"
jsonschema = "^4.23.0"
PyYAML = "^6.0.1"
PyPDF2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q"
python_files = ["test_*.py"]
